Add serial.print in balance() for the purpose of debugging
Since the result of abs(current_v - consvmin) is unknown, serial.prints
are inserted for the usage of checking the value, hence determine
whether the equations are applicable or not.

The objective next time is to test the output of the serial.prints in
balance() and then fix the issue of balance calculation equations.