Add timer implementation but keep it commented yet
In this commitment, the timer interrupt mechanism is introduced.
However, the final version is not determined yet so the current
timer setup is commented

On the other hand, the work() and charge() functions are independent
from the balance(), but the thresholds are also not determined yet,
so the part of the flow is not completed.

In next objective:
* Timer will be completed and deployed
* Thresholds of charge() and work() will be decided and deployed