Fix temperature monitor registers, and determined threshold voltages
The temperature detection lines has 5 in each IC, so the index for
each IC is set to 5.

P.S. The previous code with broken channels noted was because the
GPIOs on the previous 6811 was broken, so the channel was changed
to 1 ~ 8 but with 2, 5, 6 channel disabled. Fixed in the current
code.

Also, the threshold voltage is set to 0.3, which is a approximate
number of the working current * battery internal resistance.

The charging balance function's threshold is also set to 0.3 for
now, may be adjusted whenever the charging power(current) is
determined.