Disable timer and fix all temperature detection problems
The timer cannot work properly once the power is on, so the
solution of timer is abandomed and the traditional void loop()
returned.

Also, the temperature detection issues is fixed by changing the
array of status to determine the avalibility of gpio to all true,
which is kind of dumb that we didn't notice at first point. Anyway,
it is solved