Add charging condition check function and remove SD write function
In order to monitor the condition of charging, a table recording
whether the cells is charged to 4.12V is set and once over 90
percent of the cells is charged over 4.12V, BMS will enter FAULT
status to stop charging.

Also, since both LTC6811 and SD card use SPI to communicate with
master board, so if SD card is plugged in, master board wouldn't be
able to read any information from slaves, so all SD relative fucntions
are disabled for now. Maybe the best way is still to implement
communication through CAN bus, which is still currently developing.