Tune the charging flow in more detail
Since the threshold in balance function which is 0.2 is too big for
over charging detection, so the overcharge discharge flow is additionally
added in the charge_detect function. The objective is to discharge the
cells that is over 4.13V and wait till 90% if cells are charged over
4.12V. While balancing function still hold the threshold of 0.2V
but working among all range.