Change the sampling frequency according to judges suggestion
After the design event of FST, they suggest that we rise the frequency
of the sampling rate of our BMS monitor system and fault detection
for the sake of safty concern.

There's also other urges they suggested us to improve in the overall
strategy of balancing and Charge/Work power management. The result now
is still far away from perfect, and the rest of the suggestions will be
listed in another individual markdown document in this repo.