Add more detail on improvement list, fix print messages frequency
An equipment is found need to implement CV stage charging, which is
current sensor. Since the CV charging stage needs to set a static
voltage and keep it charging until the current decrease down to 0,
the system has to obtain the current data in order to know when the
current is small enough to indicate the battery is charged enough.
The CV stage charging strategy will be added after current sensor
is equipped.

In addition, the monitoring frequency is raised from 1Hz to 4Hz for
the sake of safty, but we don't want the data to be printed in that
rate, so a global counter is set to record the loop number, the data
will still be printed in the rate of 1Hz while monitoring remain 4Hz.